FishChatBot/requirements.txt
fastapi 
uvicorn 
langchain 
langchain-openai 
langchain_community 
mysql-connector-python
python-dotenv